+++ rc/src/encoder/encoder.c
@@ -8,13 +8,16 @@
  */
 
 #include <stdio.h>
+#include <stdint.h>
+#include <stdbool.h>
 #include <pico/stdlib.h>
 #include <hardware/pwm.h>
 #include <rc/encoder/encoder.h>
+#include <rc/defs/mbot_defs.h>
 #include "quadrature_encoder.pio.h"
 
 PIO pio;
-uint sm0, sm1, sm2;
+uint16_t sm0, sm1, sm2;
 
 int rc_encoder_init() {
     stdio_init_all();
@@ -23,7 +26,7 @@ int rc_encoder_init() {
     sm1 = pio_claim_unused_sm(pio, true);
     sm2 = pio_claim_unused_sm(pio, true);
 
-    uint offset = pio_add_program(pio, &quadrature_encoder_program);
+    uint16_t offset = pio_add_program(pio, &quadrature_encoder_program);
 
     quadrature_encoder_program_init(pio, sm0, offset, ENC0_A_PIN, 0);
     quadrature_encoder_program_init(pio, sm1, offset, ENC1_A_PIN, 0);
@@ -33,7 +36,7 @@ int rc_encoder_init() {
     quadrature_encoder_set_count(sm1, 0);
     quadrature_encoder_set_count(sm2, 0);
     
-    for(uint i = 0; i<3; i++){
+    for(uint8_t i = 0; i<3; i++){
         quadrature_encoder_request_delta(pio, sm0);
         quadrature_encoder_request_delta(pio, sm1);
         quadrature_encoder_request_delta(pio, sm2);
@@ -54,7 +57,7 @@ int rc_encoder_cleanup() {
     return (!pio_sm_is_claimed(pio, sm0) & !pio_sm_is_claimed(pio, sm1) & !pio_sm_is_claimed(pio, sm2)) ? 0 : -1;
 }
 
-int rc_encoder_read_delta(uint ch) {
+int rc_encoder_read_delta(uint8_t ch) {
     switch (ch) {
         case 0:
             return quadrature_encoder_get_delta(pio, sm0);
@@ -68,7 +71,7 @@ int rc_encoder_read_delta(uint ch) {
     }
 }
 
-int rc_encoder_read_count(uint ch) {
+int rc_encoder_read_count(uint8_t ch) {
     switch (ch) {
         case 0:
             return quadrature_encoder_get_count(sm0);
@@ -84,7 +87,7 @@ int rc_encoder_read_count(uint ch) {
 }
 
 
-int rc_encoder_write(uint ch, int pos) {
+int rc_encoder_write(uint8_t ch, int pos) {
     switch (ch) {
         case 0:
             quadrature_encoder_set_count(sm0, pos);
